Split DNAinator into main and ui classes
